LoginPage Modify
loginIcon add, loginPage layout Modify
