+++ src/App.css
@@ -37,10 +37,23 @@
   }
 }
 
+.background {
+  width: 100%;
+  height: 100%;
+  text-align: center;
+  background-color: #FAFAFA;
+  position: fixed;
+}
+
+.flex-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 .centered {
   position: fixed;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  text-align: center;
 }
