+++ src/components/index.js
@@ -1,4 +1,4 @@
-export {default as LoginModal} from './loginModal';
+export {default as LoginItem} from './loginItem';
 export {default as SignUpModal} from './signUpModal';
 export {default as TestContents} from './testContents';
 export {default as ProfileAccountItem} from './profileAccountItem';
