+++ src/images/index.js
@@ -1,1 +1,2 @@
 export {default as loginPic} from './login.png';
+export {default as loginIcon} from './loginIcon1.png';
